Initial file commit
Nothing magical here, all I did was create a new react app with typescript, then I simply uploaded (and will continue to add my files) the /src directory. The App.tsx file is slightly cluttered, all of the imported components are still there and in the returned JSX, that's OK. The code is commented is most cases and as of this initial version, everything works and is functional.